# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Campus Second-hand Trading Platform API
  version: 0.1.0
servers:
  - url: http://localhost:7001
paths:
...
  /api/items:
    get:
      operationId: listItems
      summary: 获取商品列表（分页，可选分类筛选）
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: 页码，从 1 开始
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: 每页数量
        - name: category
          in: query
          schema:
            type: string
          description: 分类筛选（可选，传"全部"或不传则不筛选）
      responses:
        "200":
          description: 商品列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemListResponse"
    post:
      operationId: createItem
      summary: 发布商品（需认证）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemRequest"
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Item"
        "400":
          description: 输入验证失败
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/items/{id}:
    get:
      operationId: getItemById
      summary: 获取商品详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 商品 ID
      responses:
        "200":
          description: 商品详情
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Item"
        "400":
          description: 无效的商品 ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 商品不存在或已下架
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateItem
      summary: 编辑商品（需认证，仅拥有者）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItemInput"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Item"
        "400":
          description: 无效的输入
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 无权修改此商品
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 商品不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/sellers/{id}:
    get:
      operationId: getSellerInfo
      summary: 获取商家基本信息
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 商家 ID
      responses:
        "200":
          description: 商家信息
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/SellerInfo"
        "400":
          description: 无效的商家 ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 商家不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/sellers/{id}/items:
    get:
      operationId: listSellerItems
      summary: 获取商家商品列表（仅含在售商品）
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 商家 ID
        - name: category
          in: query
          schema:
            type: string
          description: 分类筛选（可选，传"全部"或不传则不筛选）
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: 页码，从 1 开始
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: 每页数量
      responses:
        "200":
          description: 商品列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemListResponse"
        "400":
          description: 无效的商家 ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 商家不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/favorites/items:
    get:
      operationId: listFavoriteItems
      summary: 获取收藏商品列表（需认证）
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: 页码，从 1 开始
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: 每页数量
      responses:
        "200":
          description: 收藏商品列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteItemListResponse"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/favorites/items/{id}:
    post:
      operationId: toggleFavoriteItem
      summary: 收藏商品（切换式，需认证）
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 商品 ID
      responses:
        "200":
          description: 收藏切换成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToggleFavoriteResponse"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: unfavoriteItem
      summary: 取消收藏商品（需认证）
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 商品 ID
      responses:
        "200":
          description: 取消收藏成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/favorites/sellers:
    get:
      operationId: listFavoriteSellers
      summary: 获取收藏商家列表（需认证）
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: 页码，从 1 开始
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: 每页数量
      responses:
        "200":
          description: 收藏商家列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteSellerListResponse"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/favorites/sellers/{id}:
    post:
      operationId: toggleFavoriteSeller
      summary: 收藏商家（切换式，需认证）
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 商家 ID
      responses:
        "200":
          description: 收藏切换成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToggleFavoriteResponse"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: unfavoriteSeller
      summary: 取消收藏商家（需认证）
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 商家 ID
      responses:
        "200":
          description: 取消收藏成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/orders:
    get:
      operationId: listOrders
      summary: 查看我的购买订单（需认证）
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 订单列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/orders/sales:
    get:
      operationId: listSales
      summary: 查看售出订单（需认证）
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 售出订单列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/orders/{id}/receive:
    post:
      operationId: confirmReceive
      summary: 确认签收（需认证，仅买家可操作）
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 已确认签收
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "400":
          description: 订单状态错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 无权操作此订单
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: hex
  schemas:
...
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    Item:
      type: object
      required:
        [
          id,
          title,
          price,
          quantity,
          description,
          images,
          coverImage,
          category,
          sellerId,
          sellerName,
          status,
          createdAt,
        ]
      properties:
        id:
          type: integer
        title:
          type: string
        price:
          type: number
          format: float
        quantity:
          type: integer
        description:
          type: string
        images:
          type: array
          items:
            type: string
        coverImage:
          type: string
        category:
          type: string
        sellerId:
          type: integer
        sellerName:
          type: string
        status:
          type: string
          enum: [pending, active, delisted]
        createdAt:
          type: string
          format: date-time
        quantityUpdatedAt:
          type: string
          format: date-time
          nullable: true
        isItemFavorited:
          type: boolean
          description: 当前用户是否收藏该商品（仅登录用户返回有效值）
        isSellerFavorited:
          type: boolean
          description: 当前用户是否收藏该商家（仅登录用户返回有效值）
    ItemListResponse:
      type: object
      required: [data, total, totalPages]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        total:
          type: integer
        totalPages:
          type: integer
    CreateItemRequest:
      type: object
      additionalProperties: false
      required: [title, price, quantity, category]
      properties:
        title:
          type: string
          description: 商品名称
        price:
          type: number
          format: float
          exclusiveMinimum: 0
          description: 商品价格，必须大于 0
        quantity:
          type: integer
          minimum: 0
          description: 商品数量，必须为非负整数
        description:
          type: string
          description: 商品描述（可选）
        category:
          type: string
          description: 商品分类，必须是合法分类之一
        images:
          type: array
          items:
            type: string
          description: 图片 URL 列表（可选）
    UpdateItemInput:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          description: 商品名称
        price:
          type: number
          format: float
          exclusiveMinimum: 0
          description: 商品价格，必须大于 0
        quantity:
          type: integer
          minimum: 0
          description: 商品数量，必须为非负整数
        description:
          type: string
          description: 商品描述
        category:
          type: string
          description: 商品分类
        images:
          type: array
          items:
            type: string
          description: 图片 URL 列表
        coverImage:
          type: string
          description: 封面图片 URL
...
    SellerInfo:
      type: object
      required: [id, username, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        createdAt:
          type: string
          format: date-time
    ToggleFavoriteResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [action]
          properties:
            action:
              type: string
              enum: [favorited, unfavorited]
    FavoriteItemWithInfo:
      type: object
      required: [id, userId, itemId, createdAt, item]
      properties:
        id:
          type: integer
        userId:
          type: integer
        itemId:
          type: integer
        createdAt:
          type: string
          format: date-time
        item:
          $ref: "#/components/schemas/Item"
    FavoriteItemListResponse:
      type: object
      required: [data, total, totalPages]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/FavoriteItemWithInfo"
        total:
          type: integer
        totalPages:
          type: integer
    FavoriteSellerWithInfo:
      type: object
      required: [id, userId, sellerId, createdAt, sellerName, activeItemCount]
      properties:
        id:
          type: integer
        userId:
          type: integer
        sellerId:
          type: integer
        createdAt:
          type: string
          format: date-time
        sellerName:
          type: string
        activeItemCount:
          type: integer
    FavoriteSellerListResponse:
      type: object
      required: [data, total, totalPages]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/FavoriteSellerWithInfo"
        total:
          type: integer
        totalPages:
          type: integer
...
    OrderItem:
      type: object
      required:
        [id, orderId, itemId, sellerId, title, price, quantity, coverImage]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        itemId:
          type: integer
        sellerId:
          type: integer
        title:
          type: string
        price:
          type: number
        quantity:
          type: integer
        coverImage:
          type: string
    Order:
      type: object
      required: [id, userId, totalPrice, status, createdAt, items]
      properties:
        id:
          type: integer
        userId:
          type: integer
        totalPrice:
          type: number
        status:
          type: string
          enum: [pending_receipt, received]
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderListResponse:
      type: object
      required: [data, total, totalPages]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        totalPages:
          type: integer
...